完成读者附件excel表格的上传

+++ LibrarySystem/src/com/cc/library/action/FileUploadAction.java
@@ -35,25 +35,33 @@ public class FileUploadAction {
     
     
 	public void fileUpload() throws IOException{
-        //设置上传的路径
-        String path = ServletActionContext.getServletContext().getRealPath("/ProductImg");
+		  //设置上传的路径
+        String path = ServletActionContext.getServletContext().getRealPath("/upload");
         uploadFileName = +System.currentTimeMillis()+"_"+uploadFileName;
         File file = new File(path, uploadFileName);
-        try {
+		if(uploadContentType!=null && uploadContentType.equals("application/vnd.ms-excel")){
+			 try {
 
-            FileUtils.copyFile(upload, file);
-            upload.delete();
-            
-            //将要传入页面的数据变成json格式
+		            FileUtils.copyFile(upload, file);
+		            upload.delete();
+		            
+		            //将要传入页面的数据变成json格式
+		            JSONObject jsonObject=new JSONObject();
+		            String path2 = "upload/"+uploadFileName;
+		            jsonObject.put("path", path2);
+		            jsonObject.put("state", "success");
+		            //控制台输出格式化的json数据
+		            
+		            ServletActionContext.getResponse().getWriter().print(jsonObject);
+		        } catch (IOException e) {
+		            e.printStackTrace();
+		        }
+		}else{
+			 //将要传入页面的数据变成json格式
             JSONObject jsonObject=new JSONObject();
-            String path2 = "ProductImg/"+uploadFileName;
-            jsonObject.put("path", path2);
-            //控制台输出格式化的json数据
-            
+            jsonObject.put("state", "error");
             ServletActionContext.getResponse().getWriter().print(jsonObject);
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
+		}
